+++ facade/src/main/scala/react/semanticui/elements/placeholder/package.scala
@@ -1,13 +1,24 @@
 package react.semanticui.elements
 
-import react.common.EnumValueB
+import react.common.EnumValue
 
 package placeholder {
-  sealed trait LoaderInline extends Product with Serializable
-  object LoaderInline {
-    implicit val enum: EnumValueB[LoaderInline] = EnumValueB.toLowerCaseStringT(Inline)
-    case object Inline extends LoaderInline
-    case object Centered extends LoaderInline
+  sealed trait LineLength extends Product with Serializable
+  object LineLength {
+    implicit val enum: EnumValue[LineLength] = EnumValue.instance {
+      case Full      => "full"
+      case VeryLong  => "very long"
+      case Long      => "long"
+      case Medium    => "medium"
+      case Short     => "short"
+      case VeryShort => "very short"
+    }
+    case object Full extends LineLength
+    case object VeryLong extends LineLength
+    case object Long extends LineLength
+    case object Medium extends LineLength
+    case object Short extends LineLength
+    case object VeryShort extends LineLength
   }
 }